backend/requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
deprecation==2.1.0
distro==1.9.0
docopt==0.6.2
et_xmlfile==2.0.0
fastapi==0.121.0
Flask==3.1.2
flask-cors==6.0.1
ftfy==6.3.1
gensim==4.4.0
google-auth==2.43.0
google-genai==1.52.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.1
lightgbm==4.6.0
MarkupSafe==3.0.3
multidict==6.7.0
numpy==2.3.2
openai==1.102.0
opencv-python==4.11.0.86
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
# pdf2image==1.17.0
pillow==12.0.0
pipreqs==0.4.13
postgrest==2.24.0
propcache==0.4.1
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==2.10.1
pypdf==6.1.3
PyPDF2==3.0.1
PySocks==1.7.1
pytesseract==0.3.13
pythainlp==5.0.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
RapidFuzz==3.14.0
realtime==2.24.0
requests==2.32.5
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.16.3
selenium==4.35.0
six==1.17.0
smart_open==7.4.4
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.49.3
storage3==2.24.0
StrEnum==0.4.15
supabase==2.24.0
supabase-auth==2.24.0
supabase-functions==2.24.0
tenacity==9.1.2
threadpoolctl==3.6.0
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typhoon-ocr==0.3.8
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==2.0.0
wsproto==1.2.0
yarg==0.1.10
yarl==1.22.0
pypdfium2 
